+++ CHANGELOG.md
@@ -22,7 +22,8 @@ Changelogs for this project are recorded in this file since v0.2.0.
 
 ### Changed
 
-* The `shapelets` module now depends on Keras3+ (should be keras-backend-blind) and not anymore on TF
+* The `shapelets` module now depends on Keras3+ and Pytorch and not anymore on TF
+* Fixed `LearningShapelets` with variable length timeseries 
 
 ## [v0.6.4]